pyproject.toml
[tool.poetry]
name = "helloadventurer"
version = "0.0.1"
description = ""
authors = [ "mateusz" ]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
llama-cpp-python = "^0.2.19"
jinja2 = "^3.1.2"
langchain = "^0.0.340"